+++ v_ase/__init__.py
@@ -8,7 +8,7 @@ from v_ase.viewer import ASEEditor, view, view_edit, view_file
 try:
     __version__ = version("v_ase-gui")
 except PackageNotFoundError:
-    __version__ = "0.0.70"
+    __version__ = "0.0.71"
 
 __all__ = [
     "ASEEditor",